Pruebas y correcciones sobre el alta, eliminación y listado de artícu…

+++ src/main/java/edu/egg/loquebusques/servicios/ArticuloServicio.java
@@ -89,8 +89,15 @@ public class ArticuloServicio {
     @Transactional
     public void eliminarPorId(Integer id) {
         Articulo articulo = articuloRepositorio.findById(id).get();
-        demoraRepositorio.deleteById(articulo.getDemora().getId());
+        if(articulo.getDemora()!=null){
+            demoraRepositorio.deleteById(articulo.getDemora().getId());
+        }
         articuloRepositorio.deleteById(id);
+    }
+
+    @Transactional(readOnly = true)
+    public List<Articulo> articulosDeUnEmprendimiento(Integer emprendimiento_id) {
+        return articuloRepositorio.articulosDeUnEmprendimiento(emprendimiento_id);
     }
 
 }
